Move ChecksumValue into a separate file
