+++ common/src/main/scala/uk/ac/wellcome/platform/archive/common/verify/Checksum.scala
@@ -1,11 +1,8 @@
 package uk.ac.wellcome.platform.archive.common.verify
 
-import java.io.InputStream
-
-import org.scanamo.DynamoFormat
 import grizzled.slf4j.Logging
-import io.circe.{Decoder, Encoder, HCursor, Json}
 
+import java.io.InputStream
 import scala.util.Try
 
 case class Checksum(
@@ -30,25 +27,4 @@ case object Checksum extends Logging {
     debug(s"Got: $checksum")
     checksum
   }
-}
-
-case class ChecksumValue(value: String) {
-  override def toString: String = value
-}
-
-object ChecksumValue extends Logging {
-  def create(raw: String): ChecksumValue =
-    ChecksumValue(raw.trim)
-
-  implicit val encoder: Encoder[ChecksumValue] = (value: ChecksumValue) =>
-    Json.fromString(value.toString)
-
-  implicit val decoder: Decoder[ChecksumValue] = (cursor: HCursor) =>
-    cursor.value.as[String].map(ChecksumValue(_))
-
-  implicit def format: DynamoFormat[ChecksumValue] =
-    DynamoFormat.iso[ChecksumValue, String](
-      ChecksumValue(_),
-      _.toString
-    )
 }
